refactor: CollectionOfTheMonthCandidatesTable
